+++ src/bootstrap.php
@@ -1,5 +1,0 @@
-<?php
-
-// For multibyte string functions.
-// @see http://php.net/manual/en/book.mbstring.php
-mb_internal_encoding('UTF-8');